backend_platform/requirements.backend.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.12.0
joblib>=1.3.2
prometheus-client>=0.20.0
python-json-logger>=2.0.7
cachetools>=5.3.3
orjson>=3.10.0
onnxruntime>=1.17.0
pyyaml>=6.0.1
